Use uint32_t for num_ports (and poly), matches LV2 and size_t is excessive on 64-bit. Remove (linear) sorted assertion from Table, except in unit tests.
git-svn-id: http://svn.drobilla.net/lad/ingen@643 a436a847-0d15-0410-975c-d299462d15a1

+++ src/libs/client/DeprecatedLoader.cpp
@@ -185,7 +185,7 @@ string
 DeprecatedLoader::load_patch(const Glib::ustring&  filename,
                              boost::optional<Path> parent_path,
                              string                name,
-                             size_t                poly,
+                             uint32_t              poly,
                              MetadataMap           initial_data,
                              bool                  existing)
 {
