+++ src/libs/engine/DuplexPort.hpp
@@ -41,7 +41,7 @@ class Node;
 class DuplexPort : public InputPort, public OutputPort
 {
 public:
-	DuplexPort(Node* parent, const std::string& name, size_t index, size_t poly, DataType type, size_t buffer_size, bool is_output);
+	DuplexPort(Node* parent, const std::string& name, uint32_t index, uint32_t poly, DataType type, size_t buffer_size, bool is_output);
 	virtual ~DuplexPort() {}
 
 	void pre_process(SampleCount nframes, FrameTime start, FrameTime end);
