+++ src/libs/engine/InputPort.hpp
@@ -47,7 +47,7 @@ class Node;
 class InputPort : virtual public Port
 {
 public:
-	InputPort(Node* parent, const string& name, size_t index, size_t poly, DataType type, size_t buffer_size);
+	InputPort(Node* parent, const string& name, uint32_t index, uint32_t poly, DataType type, size_t buffer_size);
 	virtual ~InputPort() {}
 	
 	void                         add_connection(Raul::ListNode<Connection*>* c);
