+++ src/libs/engine/MidiTriggerNode.hpp
@@ -43,7 +43,7 @@ class OutputPort;
 class MidiTriggerNode : public InternalNode
 {
 public:
-	MidiTriggerNode(const string& path, size_t poly, Patch* parent, SampleRate srate, size_t buffer_size);
+	MidiTriggerNode(const string& path, uint32_t poly, Patch* parent, SampleRate srate, size_t buffer_size);
 
 	void process(SampleCount nframes, FrameTime start, FrameTime end);
